password_manager/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-WTF==1.0.0
Flask-Login==0.5.0
Flask-Talisman==1.0.0
requests==2.28.1
cryptography==39.0.1
python-dotenv==0.21.0
validators==0.20.0
gunicorn==20.1.0
cx_Oracle